Fix bitsandbytes imports to avoid ImportErrors on MacOS.
