+++ tests/backend/model_manager/load/model_cache/torch_module_autocast/test_autocast_modules.py
@@ -1,12 +1,17 @@
 import pytest
 import torch
 
-from invokeai.backend.model_manager.load.model_cache.torch_module_autocast.autocast_modules import (
-    CustomInvokeLinear8bitLt,
-    CustomInvokeLinearNF4,
-)
-from invokeai.backend.quantization.bnb_llm_int8 import InvokeLinear8bitLt
-from invokeai.backend.quantization.bnb_nf4 import InvokeLinearNF4
+if not torch.cuda.is_available():
+    pytest.skip("CUDA is not available", allow_module_level=True)
+else:
+    from invokeai.backend.model_manager.load.model_cache.torch_module_autocast.custom_invoke_linear_8_bit_lt import (
+        CustomInvokeLinear8bitLt,
+    )
+    from invokeai.backend.model_manager.load.model_cache.torch_module_autocast.custom_invoke_linear_nf4 import (
+        CustomInvokeLinearNF4,
+    )
+    from invokeai.backend.quantization.bnb_llm_int8 import InvokeLinear8bitLt
+    from invokeai.backend.quantization.bnb_nf4 import InvokeLinearNF4
 
 
 @pytest.fixture
